SMAL-148 new view for CSO upload guidance (#45)
* SAML-117 - Add producer size enum

Added producer size enum and updated RegistrationApplicationSession to have nullable producer size property. Updated the registration application service to take this value and store in session.

* SMAL-117 - Update session with orgname

this sets the organisation name based on the producer size and whether or not it is a CSO

* SMAL-117 - Set organisation name from session if not null

* SMAL-117 - Read producer size from query string

* SMAL-117 - Slight change to introduce caption on viewmodel

* SMAL-117 - small change

* SMAL-117 - only set caption if CSO

* SMAL-117 - Undo some refactoring

* SMAL-117 - Remove comments

* SMAL-117 - Small update to front end

* small test

* signed commits

* SMAL-148 - Change to be a bool for showing caption

Change so that the caption is shown if this flag is set, rather than have the text of the caption set on the view model.

* SMAL-148 - Update task list resx title for CSO

* SMAL-148 - Add resource file for new view

* SMAL-148 - Add new view for CSO upload guidance

* SMAL-148 - Correct localisation heading

* SMAL-148 - Add missing translations

* SMAL-148 - Sonar issue fix